+++ lib/src/classes/general/EzWarning.dart
