+++ lib/src/classes/screen-reader-compliance/EzImage.dart
