+++ lib/src/classes/screen-reader-compliance/EzLinkImage.dart
