+++ lib/src/classes/screen-reader-compliance/EzNewLine.dart
