+++ lib/src/classes/user-customization/EzRow.dart
