+++ lib/src/classes/user-customization/EzThemeData.dart
@@ -3,7 +3,7 @@
  * See LICENSE for distribution and usage details.
  */
 
-import '../../empathetech_flutter_ui.dart';
+import '../../../empathetech_flutter_ui.dart';
 
 import 'package:flutter/material.dart';
