+++ lib/src/consts/EzConfigValues.dart
