+++ lib/src/consts/empathetechConsts.dart
@@ -5,12 +5,8 @@
 
 // General //
 
-const String homeRoute = '/';
 const String efui = "Empathetech Flutter UI";
 const String efuiFix = "Empathetic Flutter UI";
-
-/// Allows for app images whose default is an asset to be "null"
-const String noImageValue = 'noImage';
 
 // Documentation links //
