--- lib/src/functions/textStyle.dart
@@ -1,81 +1,0 @@
-/* empathetech_flutter_ui
- * Copyright (c) 2023 Empathetech LLC. All rights reserved.
- * See LICENSE for distribution and usage details.
- */
-
-import 'package:flutter/material.dart';
-
-/// Returns the [TextTheme.displayLarge] of the current [context]
-TextStyle? displayLarge(BuildContext context) {
-  return Theme.of(context).textTheme.displayLarge;
-}
-
-/// Returns the [TextTheme.displayMedium] of the current [context]
-TextStyle? displayMedium(BuildContext context) {
-  return Theme.of(context).textTheme.displayMedium;
-}
-
-/// Returns the [TextTheme.displaySmall] of the current [context]
-TextStyle? displaySmall(BuildContext context) {
-  return Theme.of(context).textTheme.displaySmall;
-}
-
-/// Returns the [TextTheme.headlineLarge] of the current [context]
-TextStyle? headlineLarge(BuildContext context) {
-  return Theme.of(context).textTheme.headlineLarge;
-}
-
-/// Returns the [TextTheme.headlineMedium] of the current [context]
-TextStyle? headlineMedium(BuildContext context) {
-  return Theme.of(context).textTheme.headlineMedium;
-}
-
-/// Returns the [TextTheme.headlineSmall] of the current [context]
-TextStyle? headlineSmall(BuildContext context) {
-  return Theme.of(context).textTheme.headlineSmall;
-}
-
-/// Returns the [TextTheme.titleLarge] of the current [context]
-TextStyle? titleLarge(BuildContext context) {
-  return Theme.of(context).textTheme.titleLarge;
-}
-
-/// Returns the [TextTheme.titleMedium] of the current [context]
-TextStyle? titleMedium(BuildContext context) {
-  return Theme.of(context).textTheme.titleMedium;
-}
-
-/// Returns the [TextTheme.titleSmall] of the current [context]
-TextStyle? titleSmall(BuildContext context) {
-  return Theme.of(context).textTheme.titleSmall;
-}
-
-/// Returns the [TextTheme.labelLarge] of the current [context]
-TextStyle? labelLarge(BuildContext context) {
-  return Theme.of(context).textTheme.labelLarge;
-}
-
-/// Returns the [TextTheme.labelMedium] of the current [context]
-TextStyle? labelMedium(BuildContext context) {
-  return Theme.of(context).textTheme.labelMedium;
-}
-
-/// Returns the [TextTheme.labelSmall] of the current [context]
-TextStyle? labelSmall(BuildContext context) {
-  return Theme.of(context).textTheme.labelSmall;
-}
-
-/// Returns the [TextTheme.bodyLarge] of the current [context]
-TextStyle? bodyLarge(BuildContext context) {
-  return Theme.of(context).textTheme.bodyLarge;
-}
-
-/// Returns the [TextTheme.bodyMedium] of the current [context]
-TextStyle? bodyMedium(BuildContext context) {
-  return Theme.of(context).textTheme.bodyMedium;
-}
-
-/// Returns the [TextTheme.bodySmall] of the current [context]
-TextStyle? bodySmall(BuildContext context) {
-  return Theme.of(context).textTheme.bodySmall;
-}
